+++ src/main/java/kitchenpos/application/TableGroupService.java
@@ -6,9 +6,9 @@ import java.util.stream.Collectors;
 import kitchenpos.dao.OrderRepository;
 import kitchenpos.dao.OrderTableRepository;
 import kitchenpos.dao.TableGroupRepository;
-import kitchenpos.domain.OrderStatus;
-import kitchenpos.domain.OrderTable;
-import kitchenpos.domain.TableGroup;
+import kitchenpos.domain.order.OrderStatus;
+import kitchenpos.domain.table.OrderTable;
+import kitchenpos.domain.table.TableGroup;
 import kitchenpos.ui.dto.TableGroupCreateRequest;
 import kitchenpos.ui.dto.TableGroupResponse;
 import kitchenpos.ui.dto.TableResponse;
