+++ src/main/java/kitchenpos/domain/order/OrderLineItem.java
@@ -1,4 +1,4 @@
-package kitchenpos.domain;
+package kitchenpos.domain.order;
 
 import java.util.Objects;
 import javax.persistence.Embedded;
