+++ src/test/java/kitchenpos/domain/MenuProductTest.java
@@ -1,11 +1,10 @@
 package kitchenpos.domain;
 
 import static org.assertj.core.api.Assertions.assertThat;
-import static org.junit.jupiter.api.Assertions.*;
 
 import java.math.BigDecimal;
+import kitchenpos.domain.menu.MenuProduct;
 import kitchenpos.fixtures.TestFixtures;
-import org.junit.jupiter.api.Test;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.ValueSource;
