+++ src/test/java/kitchenpos/domain/OrderTableTest.java
@@ -8,6 +8,8 @@ import static org.junit.jupiter.api.Assertions.assertAll;
 
 import java.time.LocalDateTime;
 import java.util.Collections;
+import kitchenpos.domain.table.OrderTable;
+import kitchenpos.domain.table.TableGroup;
 import org.junit.jupiter.api.Test;
 
 class OrderTableTest {
